chore: moved toggleWishlist and toggleCart to proper folders; refactor: removed unused properties in AddressForm
