+++ src/components/client/UserAddress/components/Form/AddressForm.tsx
@@ -1,14 +1,7 @@
-"use client";
-
-import { Dispatch, FormEvent, SetStateAction } from "react";
-
 import { Address, TAddress } from "@/components/client/UserAddress";
 
 type AddressFormProps = Partial<TAddress> & {
   action?: (payload: FormData) => void;
-  onSubmit?: (e: FormEvent<HTMLFormElement>) => void;
-  setIsOpen: Dispatch<SetStateAction<boolean>>;
-  setAddresses: Dispatch<SetStateAction<TAddress[]>>;
 };
 
 export function AddressForm({
@@ -19,12 +12,10 @@ export function AddressForm({
   action,
   surname,
   address,
-  onSubmit,
 }: AddressFormProps) {
   return (
     <form
       action={action}
-      onSubmit={onSubmit}
       className="grid grid-cols-2 gap-4 lg:grid-cols-1 [&>label>input]:bg-accent"
     >
       <Address.Input
