+++ src/components/server/Cart/actions/toggleCartAction.ts
@@ -5,19 +5,17 @@ import { revalidateTag } from "next/cache";
 import { ItemService } from "@/services/ItemService";
 import { SelectedItemsTag } from "@/components/server/Shop";
 
-export default async function toggleWishlistAction(id: string) {
+export async function toggleCartAction(id: string) {
   const existingItem = await ItemService.getSelectedItem(id);
 
   if (existingItem) {
-    // TODO: refactor (?), 'cuz no need to duplicate itemId and params/:id
     await ItemService.updateSelectedItem(id, {
-      itemId: id,
-      isInWishlist: !existingItem.isInWishlist,
+      isInCart: !existingItem.isInCart,
     });
   } else {
     await ItemService.createSelected({
       itemId: id,
-      isInWishlist: true,
+      isInCart: true,
     });
   }
