+++ src/components/server/Cart/index.ts
@@ -1,0 +1,1 @@
+export { toggleCartAction } from "@/components/server/Cart/actions/toggleCartAction";
