+++ src/components/server/Favorites/index.ts
@@ -1,4 +1,6 @@
 export { FavoriteImage } from "@/components/server/Favorites/components/FavoriteImage";
 export { FavoriteRemove } from "@/components/server/Favorites/components/FavoriteRemove";
 export { FavoriteToCart } from "@/components/server/Favorites/components/FavoriteToCart";
+export { toggleWishlistAction } from "@/components/server/Favorites/actions/toggleWishlistAction";
+
 export { default as FavoriteItem } from "@/components/server/Favorites/components/FavoriteItem";
